+++ github3/gists/__init__.py
@@ -1,0 +1,17 @@
+"""
+github3.gists
+=============
+
+Module which contains all the gist related material.
+
+Sub-modules:
+    github3.gists.gist
+    github3.gists.file
+    github3.gists.comment
+    github3.gists.history
+
+"""
+
+from .gist import Gist
+
+__all__ = [Gist]